refactor : update directory structure
